refactor(s2): move compliance pipeline orchestration to domain CommandHandler (STA-89)
Move business orchestration (pipeline, travel rule handling, event publishing)
from ComplianceCheckApplicationService to ComplianceCheckCommandHandler in the
domain layer, following the S10 MerchantCommandHandler pattern. Application
service is now a thin DTO-mapping delegate (~40 lines).
