+++ compliance-travel-rule/compliance-travel-rule/src/test/java/com/stablecoin/payments/compliance/application/service/ComplianceCheckApplicationServiceTest.java
@@ -2,459 +2,101 @@ package com.stablecoin.payments.compliance.application.service;
 
 import com.stablecoin.payments.compliance.api.request.InitiateComplianceCheckRequest;
 import com.stablecoin.payments.compliance.api.response.ComplianceCheckResponse;
-import com.stablecoin.payments.compliance.api.response.ComplianceCheckResponse.KycResultResponse;
-import com.stablecoin.payments.compliance.api.response.ComplianceCheckResponse.SanctionsResultResponse;
-import com.stablecoin.payments.compliance.api.response.ComplianceCheckResponse.TravelRuleResponse;
 import com.stablecoin.payments.compliance.api.response.CustomerRiskProfileResponse;
 import com.stablecoin.payments.compliance.application.mapper.ComplianceCheckResponseMapper;
-import com.stablecoin.payments.compliance.domain.event.ComplianceCheckFailed;
-import com.stablecoin.payments.compliance.domain.event.ComplianceCheckPassed;
-import com.stablecoin.payments.compliance.domain.event.SanctionsHitEvent;
-import com.stablecoin.payments.compliance.domain.exception.CheckNotFoundException;
-import com.stablecoin.payments.compliance.domain.exception.CustomerNotFoundException;
-import com.stablecoin.payments.compliance.domain.exception.DuplicatePaymentException;
-import com.stablecoin.payments.compliance.domain.model.ComplianceCheck;
-import com.stablecoin.payments.compliance.domain.model.RiskScoringWeights;
-import com.stablecoin.payments.compliance.domain.port.AmlProvider;
-import com.stablecoin.payments.compliance.domain.port.ComplianceCheckRepository;
-import com.stablecoin.payments.compliance.domain.port.CustomerRiskProfileRepository;
-import com.stablecoin.payments.compliance.domain.port.EventPublisher;
-import com.stablecoin.payments.compliance.domain.port.KycProvider;
-import com.stablecoin.payments.compliance.domain.port.SanctionsProvider;
-import com.stablecoin.payments.compliance.domain.port.TravelRuleProvider;
-import com.stablecoin.payments.compliance.domain.service.ComplianceCheckService;
-import com.stablecoin.payments.compliance.domain.service.RiskScoringService;
-import org.junit.jupiter.api.BeforeEach;
+import com.stablecoin.payments.compliance.domain.model.Money;
+import com.stablecoin.payments.compliance.domain.service.ComplianceCheckCommandHandler;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.math.BigDecimal;
-import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
-import static com.stablecoin.payments.compliance.fixtures.ComplianceCheckFixtures.aKycRejectedResult;
-import static com.stablecoin.payments.compliance.fixtures.ComplianceCheckFixtures.aKycResult;
-import static com.stablecoin.payments.compliance.fixtures.ComplianceCheckFixtures.aSanctionsClearResult;
-import static com.stablecoin.payments.compliance.fixtures.ComplianceCheckFixtures.aSanctionsHitResult;
-import static com.stablecoin.payments.compliance.fixtures.ComplianceCheckFixtures.anAmlClearResult;
-import static com.stablecoin.payments.compliance.fixtures.ComplianceCheckFixtures.anAmlFlaggedResult;
+import static com.stablecoin.payments.compliance.fixtures.ComplianceCheckFixtures.aPendingCheck;
 import static com.stablecoin.payments.compliance.fixtures.CustomerRiskProfileFixtures.aRiskProfile;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
 
 @ExtendWith(MockitoExtension.class)
 class ComplianceCheckApplicationServiceTest {
 
-    @Mock private ComplianceCheckRepository checkRepository;
-    @Mock private CustomerRiskProfileRepository profileRepository;
-    @Mock private KycProvider kycProvider;
-    @Mock private SanctionsProvider sanctionsProvider;
-    @Mock private AmlProvider amlProvider;
-    @Mock private TravelRuleProvider travelRuleProvider;
-    @Mock private EventPublisher<Object> eventPublisher;
+    @Mock private ComplianceCheckCommandHandler commandHandler;
+    @Mock private ComplianceCheckResponseMapper responseMapper;
 
-    private ComplianceCheckService complianceCheckService;
-    private RiskScoringService riskScoringService;
-    private ComplianceCheckResponseMapper responseMapper;
+    @InjectMocks
     private ComplianceCheckApplicationService service;
 
-    @BeforeEach
-    void setUp() {
-        complianceCheckService = new ComplianceCheckService();
-        riskScoringService = new RiskScoringService(RiskScoringWeights.defaults());
-        responseMapper = new ComplianceCheckResponseMapper();
-        service = new ComplianceCheckApplicationService(
-                checkRepository, profileRepository,
-                kycProvider, sanctionsProvider, amlProvider, travelRuleProvider,
-                complianceCheckService, riskScoringService,
-                eventPublisher, responseMapper);
-    }
-
-    private static InitiateComplianceCheckRequest aValidRequest() {
-        return new InitiateComplianceCheckRequest(
+    @Test
+    @DisplayName("initiateCheck should delegate to commandHandler and map response")
+    void initiateCheckShouldDelegateAndMap() {
+        var request = new InitiateComplianceCheckRequest(
                 UUID.randomUUID(), UUID.randomUUID(), UUID.randomUUID(),
                 new BigDecimal("1000.00"), "USD", "US", "DE", "EUR");
+        var domainCheck = aPendingCheck();
+        var expectedResponse = new ComplianceCheckResponse(
+                domainCheck.checkId(), domainCheck.paymentId(), "PENDING",
+                null, null, null, null, null, null, null,
+                domainCheck.createdAt(), null);
+
+        given(commandHandler.initiateCheck(
+                request.paymentId(), request.senderId(), request.recipientId(),
+                new Money(request.amount(), request.currency()),
+                request.sourceCountry(), request.targetCountry(), request.targetCurrency()))
+                .willReturn(domainCheck);
+        given(responseMapper.toResponse(domainCheck)).willReturn(expectedResponse);
+
+        var result = service.initiateCheck(request);
+
+        assertThat(result)
+                .usingRecursiveComparison()
+                .isEqualTo(expectedResponse);
     }
 
-    private static InitiateComplianceCheckRequest aBelowThresholdRequest() {
-        return new InitiateComplianceCheckRequest(
-                UUID.randomUUID(), UUID.randomUUID(), UUID.randomUUID(),
-                new BigDecimal("500.00"), "USD", "US", "DE", "EUR");
+    @Test
+    @DisplayName("getCheck should delegate to commandHandler and map response")
+    void getCheckShouldDelegateAndMap() {
+        var checkId = UUID.randomUUID();
+        var domainCheck = aPendingCheck();
+        var expectedResponse = new ComplianceCheckResponse(
+                domainCheck.checkId(), domainCheck.paymentId(), "PENDING",
+                null, null, null, null, null, null, null,
+                domainCheck.createdAt(), null);
+
+        given(commandHandler.getCheck(checkId)).willReturn(domainCheck);
+        given(responseMapper.toResponse(domainCheck)).willReturn(expectedResponse);
+
+        var result = service.getCheck(checkId);
+
+        assertThat(result)
+                .usingRecursiveComparison()
+                .isEqualTo(expectedResponse);
+        then(commandHandler).should().getCheck(checkId);
     }
 
-    @Nested
-    @DisplayName("Happy path — full pipeline")
-    class HappyPath {
+    @Test
+    @DisplayName("getCustomerRiskProfile should delegate to commandHandler and map response")
+    void getCustomerRiskProfileShouldDelegateAndMap() {
+        var customerId = UUID.randomUUID();
+        var profile = aRiskProfile().toBuilder().customerId(customerId).build();
+        var expectedResponse = new CustomerRiskProfileResponse(
+                customerId, "KYC_TIER_2", profile.kycVerifiedAt(),
+                "LOW", 20,
+                profile.perTxnLimitUsd(), profile.dailyLimitUsd(),
+                profile.monthlyLimitUsd(), profile.lastScoredAt());
 
-        @Test
-        @DisplayName("should complete full pipeline and return PASSED")
-        void shouldCompleteFullPipeline() {
-            var request = aValidRequest();
-            given(checkRepository.findByPaymentId(request.paymentId())).willReturn(Optional.empty());
-            given(kycProvider.verify(request.senderId(), request.recipientId()))
-                    .willReturn(aKycResult(null));
-            given(sanctionsProvider.screen(request.senderId(), request.recipientId()))
-                    .willReturn(aSanctionsClearResult(null));
-            given(amlProvider.analyze(request.senderId(), request.recipientId()))
-                    .willReturn(anAmlClearResult(null));
-            given(profileRepository.findByCustomerId(request.senderId()))
-                    .willReturn(Optional.of(aRiskProfile()));
-            given(travelRuleProvider.transmit(any())).willReturn("tr-ref-123");
-            given(checkRepository.save(any(ComplianceCheck.class)))
-                    .willAnswer(invocation -> invocation.getArgument(0));
+        given(commandHandler.getCustomerRiskProfile(customerId)).willReturn(profile);
+        given(responseMapper.toResponse(profile)).willReturn(expectedResponse);
 
-            var response = service.initiateCheck(request);
+        var result = service.getCustomerRiskProfile(customerId);
 
-            var expected = new ComplianceCheckResponse(
-                    null, request.paymentId(), "PASSED", "PASSED",
-                    null,
-                    new KycResultResponse("VERIFIED", "VERIFIED", "KYC_TIER_2"),
-                    new SanctionsResultResponse(false, false, List.of("OFAC", "EU", "UN")),
-                    new TravelRuleResponse("IVMS101", "TRANSMITTED"),
-                    null, null, null, null);
-
-            assertThat(response)
-                    .usingRecursiveComparison()
-                    .ignoringFields("checkId", "createdAt", "completedAt", "riskScore",
-                            "errorCode", "errorMessage")
-                    .isEqualTo(expected);
-        }
-
-        @Test
-        @DisplayName("should publish ComplianceCheckPassed event")
-        void shouldPublishPassedEvent() {
-            var request = aValidRequest();
-            given(checkRepository.findByPaymentId(request.paymentId())).willReturn(Optional.empty());
-            given(kycProvider.verify(any(), any())).willReturn(aKycResult(null));
-            given(sanctionsProvider.screen(any(), any())).willReturn(aSanctionsClearResult(null));
-            given(amlProvider.analyze(any(), any())).willReturn(anAmlClearResult(null));
-            given(profileRepository.findByCustomerId(any())).willReturn(Optional.of(aRiskProfile()));
-            given(travelRuleProvider.transmit(any())).willReturn("tr-ref-123");
-            given(checkRepository.save(any(ComplianceCheck.class)))
-                    .willAnswer(invocation -> invocation.getArgument(0));
-
-            service.initiateCheck(request);
-
-            var captor = ArgumentCaptor.forClass(Object.class);
-            then(eventPublisher).should().publish(captor.capture());
-
-            var expected = new ComplianceCheckPassed(
-                    null, request.paymentId(), null, 0, null, null, null);
-
-            assertThat(captor.getValue())
-                    .usingRecursiveComparison()
-                    .ignoringFields("checkId", "correlationId", "riskScore", "riskBand",
-                            "travelRuleRef", "passedAt")
-                    .isEqualTo(expected);
-        }
-
-        @Test
-        @DisplayName("should skip travel rule for below-threshold amounts")
-        void shouldSkipTravelRuleBelowThreshold() {
-            var request = aBelowThresholdRequest();
-            given(checkRepository.findByPaymentId(request.paymentId())).willReturn(Optional.empty());
-            given(kycProvider.verify(any(), any())).willReturn(aKycResult(null));
-            given(sanctionsProvider.screen(any(), any())).willReturn(aSanctionsClearResult(null));
-            given(amlProvider.analyze(any(), any())).willReturn(anAmlClearResult(null));
-            given(profileRepository.findByCustomerId(any())).willReturn(Optional.of(aRiskProfile()));
-            given(checkRepository.save(any(ComplianceCheck.class)))
-                    .willAnswer(invocation -> invocation.getArgument(0));
-
-            var response = service.initiateCheck(request);
-
-            var expected = new ComplianceCheckResponse(
-                    null, request.paymentId(), "PASSED", "PASSED",
-                    null, null, null, null, null, null, null, null);
-
-            assertThat(response)
-                    .usingRecursiveComparison()
-                    .ignoringFields("checkId", "createdAt", "completedAt", "riskScore",
-                            "kycResult", "sanctionsResult", "errorCode", "errorMessage")
-                    .isEqualTo(expected);
-            then(travelRuleProvider).should(never()).transmit(any());
-        }
-    }
-
-    @Nested
-    @DisplayName("KYC failure")
-    class KycFailure {
-
-        @Test
-        @DisplayName("should stop pipeline on KYC rejection")
-        void shouldStopOnKycRejection() {
-            var request = aValidRequest();
-            given(checkRepository.findByPaymentId(request.paymentId())).willReturn(Optional.empty());
-            given(kycProvider.verify(any(), any())).willReturn(aKycRejectedResult(null));
-            given(checkRepository.save(any(ComplianceCheck.class)))
-                    .willAnswer(invocation -> invocation.getArgument(0));
-
-            var response = service.initiateCheck(request);
-
-            var expected = new ComplianceCheckResponse(
-                    null, request.paymentId(), "FAILED", "FAILED",
-                    null, null, null, null, null, null, null, null);
-
-            assertThat(response)
-                    .usingRecursiveComparison()
-                    .ignoringFields("checkId", "createdAt", "completedAt", "kycResult",
-                            "errorCode", "errorMessage")
-                    .isEqualTo(expected);
-            then(sanctionsProvider).should(never()).screen(any(), any());
-            then(amlProvider).should(never()).analyze(any(), any());
-        }
-
-        @Test
-        @DisplayName("should publish ComplianceCheckFailed event on KYC rejection")
-        void shouldPublishFailedEventOnKycRejection() {
-            var request = aValidRequest();
-            given(checkRepository.findByPaymentId(request.paymentId())).willReturn(Optional.empty());
-            given(kycProvider.verify(any(), any())).willReturn(aKycRejectedResult(null));
-            given(checkRepository.save(any(ComplianceCheck.class)))
-                    .willAnswer(invocation -> invocation.getArgument(0));
-
-            service.initiateCheck(request);
-
-            var captor = ArgumentCaptor.forClass(Object.class);
-            then(eventPublisher).should().publish(captor.capture());
-
-            var expected = new ComplianceCheckFailed(
-                    null, request.paymentId(), null, null, null, null);
-            assertThat(captor.getValue())
-                    .usingRecursiveComparison()
-                    .ignoringFields("checkId", "correlationId", "reason", "errorCode", "failedAt")
-                    .isEqualTo(expected);
-        }
-    }
-
-    @Nested
-    @DisplayName("Sanctions hit")
-    class SanctionsHitPath {
-
-        @Test
-        @DisplayName("should stop pipeline on sanctions hit")
-        void shouldStopOnSanctionsHit() {
-            var request = aValidRequest();
-            given(checkRepository.findByPaymentId(request.paymentId())).willReturn(Optional.empty());
-            given(kycProvider.verify(any(), any())).willReturn(aKycResult(null));
-            given(sanctionsProvider.screen(any(), any())).willReturn(aSanctionsHitResult(null));
-            given(checkRepository.save(any(ComplianceCheck.class)))
-                    .willAnswer(invocation -> invocation.getArgument(0));
-
-            var response = service.initiateCheck(request);
-
-            var expected = new ComplianceCheckResponse(
-                    null, request.paymentId(), "SANCTIONS_HIT", "SANCTIONS_HIT",
-                    null, null, null, null, null, null, null, null);
-
-            assertThat(response)
-                    .usingRecursiveComparison()
-                    .ignoringFields("checkId", "createdAt", "completedAt", "kycResult",
-                            "sanctionsResult", "errorCode", "errorMessage")
-                    .isEqualTo(expected);
-            then(amlProvider).should(never()).analyze(any(), any());
-        }
-
-        @Test
-        @DisplayName("should publish both ComplianceCheckFailed and SanctionsHitEvent")
-        void shouldPublishSanctionsHitEvents() {
-            var request = aValidRequest();
-            given(checkRepository.findByPaymentId(request.paymentId())).willReturn(Optional.empty());
-            given(kycProvider.verify(any(), any())).willReturn(aKycResult(null));
-            given(sanctionsProvider.screen(any(), any())).willReturn(aSanctionsHitResult(null));
-            given(checkRepository.save(any(ComplianceCheck.class)))
-                    .willAnswer(invocation -> invocation.getArgument(0));
-
-            service.initiateCheck(request);
-
-            var captor = ArgumentCaptor.forClass(Object.class);
-            then(eventPublisher).should(times(2)).publish(captor.capture());
-            var events = captor.getAllValues();
-
-            var expectedFailed = new ComplianceCheckFailed(
-                    null, request.paymentId(), null, null, null, null);
-            assertThat(events.get(0))
-                    .usingRecursiveComparison()
-                    .ignoringFields("checkId", "correlationId", "reason", "errorCode", "failedAt")
-                    .isEqualTo(expectedFailed);
-
-            var expectedHit = new SanctionsHitEvent(
-                    null, request.paymentId(), null, "SENDER", "OFAC", null, null);
-            assertThat(events.get(1))
-                    .usingRecursiveComparison()
-                    .ignoringFields("checkId", "correlationId", "hitDetails", "detectedAt")
-                    .isEqualTo(expectedHit);
-        }
-    }
-
-    @Nested
-    @DisplayName("AML flagged")
-    class AmlFlaggedPath {
-
-        @Test
-        @DisplayName("should stop pipeline on AML flag and route to MANUAL_REVIEW")
-        void shouldStopOnAmlFlag() {
-            var request = aValidRequest();
-            given(checkRepository.findByPaymentId(request.paymentId())).willReturn(Optional.empty());
-            given(kycProvider.verify(any(), any())).willReturn(aKycResult(null));
-            given(sanctionsProvider.screen(any(), any())).willReturn(aSanctionsClearResult(null));
-            given(amlProvider.analyze(any(), any())).willReturn(anAmlFlaggedResult(null));
-            given(checkRepository.save(any(ComplianceCheck.class)))
-                    .willAnswer(invocation -> invocation.getArgument(0));
-
-            var response = service.initiateCheck(request);
-
-            var expected = new ComplianceCheckResponse(
-                    null, request.paymentId(), "MANUAL_REVIEW", "MANUAL_REVIEW",
-                    null, null, null, null, null, null, null, null);
-
-            assertThat(response)
-                    .usingRecursiveComparison()
-                    .ignoringFields("checkId", "createdAt", "completedAt", "riskScore",
-                            "kycResult", "sanctionsResult", "errorCode", "errorMessage")
-                    .isEqualTo(expected);
-            then(travelRuleProvider).should(never()).transmit(any());
-        }
-    }
-
-    @Nested
-    @DisplayName("Travel rule failure")
-    class TravelRuleFailure {
-
-        @Test
-        @DisplayName("should handle travel rule transmission failure gracefully")
-        void shouldHandleTravelRuleFailure() {
-            var request = aValidRequest();
-            given(checkRepository.findByPaymentId(request.paymentId())).willReturn(Optional.empty());
-            given(kycProvider.verify(any(), any())).willReturn(aKycResult(null));
-            given(sanctionsProvider.screen(any(), any())).willReturn(aSanctionsClearResult(null));
-            given(amlProvider.analyze(any(), any())).willReturn(anAmlClearResult(null));
-            given(profileRepository.findByCustomerId(any())).willReturn(Optional.of(aRiskProfile()));
-            given(travelRuleProvider.transmit(any())).willThrow(new RuntimeException("Network error"));
-            given(checkRepository.save(any(ComplianceCheck.class)))
-                    .willAnswer(invocation -> invocation.getArgument(0));
-
-            var response = service.initiateCheck(request);
-
-            var expected = new ComplianceCheckResponse(
-                    null, request.paymentId(), "FAILED", "FAILED",
-                    null, null, null, null, null,
-                    "Travel rule transmission failed: Network error",
-                    null, null);
-
-            assertThat(response)
-                    .usingRecursiveComparison()
-                    .ignoringFields("checkId", "createdAt", "completedAt", "riskScore",
-                            "kycResult", "sanctionsResult", "travelRule", "errorCode")
-                    .isEqualTo(expected);
-        }
-    }
-
-    @Nested
-    @DisplayName("Duplicate payment")
-    class DuplicatePayment {
-
-        @Test
-        @DisplayName("should throw DuplicatePaymentException when check already exists")
-        void shouldThrowOnDuplicate() {
-            var request = aValidRequest();
-            var existingCheck = ComplianceCheck.initiate(
-                    request.paymentId(), request.senderId(), request.recipientId(),
-                    new com.stablecoin.payments.compliance.domain.model.Money(request.amount(), request.currency()),
-                    request.sourceCountry(), request.targetCountry(), request.targetCurrency());
-            given(checkRepository.findByPaymentId(request.paymentId()))
-                    .willReturn(Optional.of(existingCheck));
-
-            assertThatThrownBy(() -> service.initiateCheck(request))
-                    .isInstanceOf(DuplicatePaymentException.class)
-                    .hasMessageContaining(request.paymentId().toString());
-
-            then(kycProvider).should(never()).verify(any(), any());
-        }
-    }
-
-    @Nested
-    @DisplayName("Get check")
-    class GetCheck {
-
-        @Test
-        @DisplayName("should return check response for existing check")
-        void shouldReturnCheckResponse() {
-            var checkId = UUID.randomUUID();
-            var check = ComplianceCheck.initiate(
-                    UUID.randomUUID(), UUID.randomUUID(), UUID.randomUUID(),
-                    new com.stablecoin.payments.compliance.domain.model.Money(new BigDecimal("1000.00"), "USD"),
-                    "US", "DE", "EUR");
-            given(checkRepository.findById(checkId)).willReturn(Optional.of(check));
-
-            var response = service.getCheck(checkId);
-
-            var expected = new ComplianceCheckResponse(
-                    check.checkId(), check.paymentId(), "PENDING",
-                    null, null, null, null, null, null, null,
-                    check.createdAt(), null);
-
-            assertThat(response)
-                    .usingRecursiveComparison()
-                    .isEqualTo(expected);
-        }
-
-        @Test
-        @DisplayName("should throw CheckNotFoundException for missing check")
-        void shouldThrowOnMissingCheck() {
-            var checkId = UUID.randomUUID();
-            given(checkRepository.findById(checkId)).willReturn(Optional.empty());
-
-            assertThatThrownBy(() -> service.getCheck(checkId))
-                    .isInstanceOf(CheckNotFoundException.class)
-                    .hasMessageContaining(checkId.toString());
-        }
-    }
-
-    @Nested
-    @DisplayName("Get customer risk profile")
-    class GetCustomerProfile {
-
-        @Test
-        @DisplayName("should return customer risk profile response")
-        void shouldReturnProfileResponse() {
-            var customerId = UUID.randomUUID();
-            var profile = aRiskProfile().toBuilder().customerId(customerId).build();
-            given(profileRepository.findByCustomerId(customerId)).willReturn(Optional.of(profile));
-
-            var response = service.getCustomerRiskProfile(customerId);
-
-            var expected = new CustomerRiskProfileResponse(
-                    customerId, "KYC_TIER_2", profile.kycVerifiedAt(),
-                    "LOW", 20,
-                    profile.perTxnLimitUsd(), profile.dailyLimitUsd(),
-                    profile.monthlyLimitUsd(), profile.lastScoredAt());
-
-            assertThat(response)
-                    .usingRecursiveComparison()
-                    .withComparatorForType(BigDecimal::compareTo, BigDecimal.class)
-                    .isEqualTo(expected);
-        }
-
-        @Test
-        @DisplayName("should throw CustomerNotFoundException for missing profile")
-        void shouldThrowOnMissingProfile() {
-            var customerId = UUID.randomUUID();
-            given(profileRepository.findByCustomerId(customerId)).willReturn(Optional.empty());
-
-            assertThatThrownBy(() -> service.getCustomerRiskProfile(customerId))
-                    .isInstanceOf(CustomerNotFoundException.class)
-                    .hasMessageContaining(customerId.toString());
-        }
+        assertThat(result)
+                .usingRecursiveComparison()
+                .isEqualTo(expectedResponse);
+        then(commandHandler).should().getCustomerRiskProfile(customerId);
     }
 }
